CoinbaseMCP/requirements.txt
fastmcp>=0.1.0
requests>=2.28.0
pynacl>=1.5.0
python-dotenv>=1.0.0
